+++ runtime/hdutils.py
@@ -79,6 +79,15 @@ class MetaDataWrapped(BaseModel):
         alias="__data__",
         description="The actual data which constitutes the pandas object.",
     )
+    data_parsing_options__: dict = Field(
+        {},
+        alias="__data_parsing_options__",
+        description=(
+            "Additional options for parsing the provided data."
+            " For example, setting orient to one of the allowed values for the respective"
+            " Pandas type allows to use different json representations for the actual data."
+        ),
+    )
 
 
 def try_parse_wrapped(
@@ -86,9 +95,7 @@ def try_parse_wrapped(
     hd_wrapped_data_object: Literal["SERIES", "DATAFRAME"],
 ) -> MetaDataWrapped:
     if isinstance(data, str):
-        wrapped_data = MetaDataWrapped.parse_raw(
-            data
-        )  # model_validate_json in pydantic 2.0
+        wrapped_data = MetaDataWrapped.parse_raw(data)  # model_validate_json in pydantic 2.0
 
         if wrapped_data.hd_wrapped_data_object__ != hd_wrapped_data_object:
             msg = (
@@ -106,18 +113,20 @@ def try_parse_wrapped(
 def parse_wrapped_content(
     v: str | dict | list,
     wrapped_data_objec: Literal["SERIES", "DATAFRAME"],
-) -> tuple[str | dict | list, None | dict[str, Any]]:
+) -> tuple[str | dict | list, None | dict[str, Any], dict[str, Any]]:
     data_content: str | dict | list
     try:
         wrapped_object = try_parse_wrapped(v, wrapped_data_objec)
         parsed_metadata = wrapped_object.metadata__
         data_content = wrapped_object.data__
+        parsing_options = wrapped_object.data_parsing_options__
     except (ValidationError, TypeError) as e:
         logger.debug("Data object is not wrapped: %s", str(e))
         data_content = v
         parsed_metadata = None
+        parsing_options = {}
 
-    return data_content, parsed_metadata
+    return data_content, parsed_metadata, parsing_options
 
 
 def wrap_metadata_as_attrs(
@@ -131,18 +140,20 @@ def wrap_metadata_as_attrs(
 
 
 def parse_pandas_data_content(
-    data_content: str | dict | list, typ: Literal["series", "frame"]
+    data_content: str | dict | list, typ: Literal["series", "frame"], parsing_options: dict
 ) -> pd.DataFrame | pd.Series:
     try:
         if isinstance(data_content, str):
-            parsed_pandas_object = pd.read_json(io.StringIO(data_content), typ=typ)
+            parsed_pandas_object = pd.read_json(
+                io.StringIO(data_content), typ=typ, **parsing_options
+            )
         else:
-            parsed_pandas_object = pd.read_json(data_content, typ=typ)
+            parsed_pandas_object = pd.read_json(data_content, typ=typ, **parsing_options)
 
     except Exception:  # noqa: BLE001
         try:
             parsed_pandas_object = pd.read_json(
-                io.StringIO(json.dumps(data_content)), typ=typ
+                io.StringIO(json.dumps(data_content)), typ=typ, **parsing_options
             )
 
         except Exception as read_json_exception:  # noqa: BLE001
@@ -207,10 +218,10 @@ class PydanticPandasSeries:
             logger.error(msg)
             raise ValueError(msg)
 
-        data_content, metadata = parse_wrapped_content(v, "SERIES")
+        data_content, metadata, parsing_options = parse_wrapped_content(v, "SERIES")
 
         return wrap_metadata_as_attrs(
-            parse_pandas_data_content(data_content, "series"), metadata
+            parse_pandas_data_content(data_content, "series", parsing_options), metadata
         )
 
 
@@ -239,16 +250,14 @@ class PydanticPandasDataFrame:
         if not isinstance(
             v, str | dict | list
         ):  # need to check at runtime since we get objects from user code
-            msg = (
-                f"Got unexpected type at runtime when parsing DataFrame: {str(type(v))}"
-            )
+            msg = f"Got unexpected type at runtime when parsing DataFrame: {str(type(v))}"
             logger.error(msg)
             raise ValueError(msg)
 
-        data_content, metadata = parse_wrapped_content(v, "DATAFRAME")
+        data_content, metadata, parsing_options = parse_wrapped_content(v, "DATAFRAME")
 
         return wrap_metadata_as_attrs(
-            parse_pandas_data_content(data_content, "frame"), metadata
+            parse_pandas_data_content(data_content, "frame", parsing_options), metadata
         )
 
 
@@ -277,10 +286,10 @@ class PydanticMultiTimeseriesPandasDataFrame:
             logger.error(msg)
             raise ValueError(msg)
 
-        data_content, metadata = parse_wrapped_content(v, "DATAFRAME")
+        data_content, metadata, parsing_options = parse_wrapped_content(v, "DATAFRAME")
 
         return wrap_metadata_as_attrs(
-            parse_pandas_data_content(data_content, "frame"), metadata
+            parse_pandas_data_content(data_content, "frame", parsing_options), metadata
         )
 
     @classmethod
@@ -426,18 +435,22 @@ class AdvancedTypesOutputSerializationConfig(BaseConfig):
             "__hd_wrapped_data_object__": "SERIES",
             "__metadata__": v.attrs,
             "__data__": json.loads(
+                # double serialization7deserialization in order to serialize both NaN and NaT
+                # to null
                 v.to_json(
-                    date_format="iso"
-                )  # in order to serialize both NaN and NaT to null,
+                    date_format="iso",
+                    orient="split",
+                    # orient="split" serialization is the only way pandas keeps duplicate index
+                    #  (with possibly different values) entries for Series objects!
+                )
             ),
+            "__data_parsing_options__": {"orient": "split"},
         },
         pd.DataFrame: lambda v: {
             "__hd_wrapped_data_object__": "DATAFRAME",
             "__metadata__": v.attrs,
             "__data__": json.loads(
-                v.to_json(
-                    date_format="iso"
-                )  # in order to serialize both NaN and NaT to null
+                v.to_json(date_format="iso")  # in order to serialize both NaN and NaT to null
             ),
         },
         PydanticPandasSeries: lambda v: {
@@ -449,24 +462,20 @@ class AdvancedTypesOutputSerializationConfig(BaseConfig):
             "__hd_wrapped_data_object__": "DATAFRAME",
             "__metadata__": v.attrs,
             "__data__": json.loads(
-                v.to_json(
-                    date_format="iso"
-                )  # in order to serialize both NaN and NaT to null
+                v.to_json(date_format="iso")  # in order to serialize both NaN and NaT to null
             ),
         },
         PydanticMultiTimeseriesPandasDataFrame: lambda v: {
             "__hd_wrapped_data_object__": "DATAFRAME",
             "__metadata__": v.attrs,
-            "__data__": v.to_dict(
-                date_format="iso"
+            "__data__": json.loads(
+                v.to_json(date_format="iso")
             ),  # in order to serialize both NaN and NaT to null
         },
         np.ndarray: lambda v: v.tolist(),
         datetime.datetime: lambda v: v.isoformat(),
         UUID: lambda v: str(v),  # alternatively: v.hex
-        Figure: lambda v: json.loads(
-            json.dumps(v.to_plotly_json(), cls=PlotlyJSONEncoder)
-        ),
+        Figure: lambda v: json.loads(json.dumps(v.to_plotly_json(), cls=PlotlyJSONEncoder)),
     }
 
 
@@ -520,9 +529,7 @@ def parse_single_value_dynamically(
 
 
 def parse_value(value: Any, data_type_str: str, nullable: bool) -> Any:
-    return parse_single_value_dynamically(
-        "some_value", value, DataType(data_type_str), nullable
-    )
+    return parse_single_value_dynamically("some_value", value, DataType(data_type_str), nullable)
 
 
 def parse_default_value(component_info: dict, input_name: str) -> Any: